+++ components/molecules/ContributorHighlight/contributor-highlight-card.tsx
@@ -17,7 +17,6 @@ import { FaUserPlus } from "react-icons/fa";
 import { GrFlag } from "react-icons/gr";
 import { useSWRConfig } from "swr";
 
-import { ToastTrigger } from "lib/utils/toast-trigger";
 import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
 import GhOpenGraphImg from "../GhOpenGraphImg/gh-open-graph-img";
 import {
@@ -44,6 +43,7 @@ import {
   AlertDialogTrigger
 } from "../AlertDialog/alert-dialog";
 import { deleteHighlight } from "lib/hooks/deleteHighlight";
+import { useToast } from "lib/hooks/useToast";
 
 interface ContributorHighlightCardProps {
   title?: string;
@@ -53,6 +53,7 @@ interface ContributorHighlightCardProps {
   id: string;
   refreshCallBack?: () => void;
 }
+
 const ContributorHighlightCard = ({
   title,
   desc,
@@ -61,6 +62,7 @@ const ContributorHighlightCard = ({
   id,
   refreshCallBack
 }: ContributorHighlightCardProps) => {
+  const { toast } = useToast();
   const twitterTweet = `${title || "Open Source Highlight"} - OpenSauced from ${user}`;
   const reportSubject = `Reported Highlight ${user}: ${title}`;
   const [highlight, setHighlight] = useState({ title, desc, prLink });
@@ -86,7 +88,7 @@ const ContributorHighlightCard = ({
     const url = new URL(content).toString();
     try {
       await navigator.clipboard.writeText(url);
-      ToastTrigger({ message: "Copied to clipboard", type: "success" });
+      toast({ description: "Copied to clipboard", variant: "success" });
     } catch (error) {
       console.log(error);
     }
@@ -118,12 +120,12 @@ const ContributorHighlightCard = ({
         );
         setLoading(false);
         if (res) {
+          toast({ description: "Highlights Updated Successfully", variant: "success" });
           refreshCallBack && refreshCallBack();
-          ToastTrigger({ message: "Highlights Updated Successfully", type: "success" });
           setOpenEdit(false);
         } else {
           setLoading(false);
-          setError("An error occurred while updating!!!");
+          setError("An error occurred while updating!");
         }
       }
     } else {
@@ -136,8 +138,10 @@ const ContributorHighlightCard = ({
     const res = await deleteHighlight(id);
     setDeleteLoading(false);
     if (res !== false) {
+      toast({ description: "Highlights deleted Successfully", variant: "success" });
+
       refreshCallBack && refreshCallBack();
-      ToastTrigger({ message: "Highlights Deleted Successfully", type: "success" });
+
       setAlertOpen(false);
       setOpenEdit(false);
       setTimeout(() => {
@@ -146,7 +150,7 @@ const ContributorHighlightCard = ({
     } else {
       console.error(res);
       setAlertOpen(false);
-      ToastTrigger({ message: "An error occured!!!", type: "error" });
+      toast({ description: "An error occured!", variant: "danger" });
     }
   };